move reusable components to the shared folder
