+++ src/components/redirect_notification/SupportedNotification_Not_Installed.js
@@ -3,10 +3,10 @@ import {
   runningInChromeDesktop,
   runningInFirefoxDesktop,
 } from "../../utils/misc/browserDetection";
-import Modal from "./Modal";
-import Header from "./Header";
-import Body from "./Body";
-import Footer from "./Footer";
+import Modal from "./shared/Modal";
+import Header from "./shared/Header";
+import Body from "./shared/Body";
+import Footer from "./shared/Footer";
 
 export default function SupportedNotificationNotInstalled({
   handleCloseRedirectionModal,
