+++ src/components/redirect_notification/shared/Body.js
@@ -1,4 +1,4 @@
-import * as s from "./RedirectionNotificationModal.sc";
+import * as s from "../RedirectionNotificationModal.sc";
 
 export default function Body({ children }) {
   return <s.Body>{children}</s.Body>;
