+++ src/components/redirect_notification/shared/Footer.js
@@ -1,4 +1,4 @@
-import * as s from "./RedirectionNotificationModal.sc";
+import * as s from "../RedirectionNotificationModal.sc";
 
 export default function Footer({ children }) {
   return <s.Footer>{children}</s.Footer>;
